Add decoder

+++ MachineEmulator/RAM.cs
@@ -8,19 +8,19 @@ public class RAM
         Data = new int[RAM_SIZE];
     }
 
-    public byte this[int key] {
+    public byte this[long key] {
         get => GetValue(key);
         set => SetValue(key, value);
     }
 
-    private byte GetValue(int key) {
+    private byte GetValue(long key) {
         int dword = Data[key / 4];
-        return (byte)((dword >> ((3 - key % 4) * 8)) & 0xFF);
+        return (byte)((dword >> (int)((3 - key % 4) * 8)) & 0xFF);
     }
 
-    private void SetValue(int key, byte value) {
+    private void SetValue(long key, byte value) {
         int dwordToSet = Data[key / 4];
-        int shiftCount = ((3 - key % 4) * 8);
+        int shiftCount = ((int)(3 - key % 4) * 8);
         int mask = 0xFF << shiftCount;
         Data[key / 4] = (int)((dwordToSet & (0xFFFFFFFF ^ mask)) + (value << shiftCount));
     }
